Make conditions filters
This starts by noting the difference between queries ("How well
does this document match this query clause") and filters ("Does
this document match this query clause?") in elasticsearch. Though these
are the same conditions, we need a separate property for each.

```ts
class ThronesSearch extends Search {
  // ... code ...
  filters!: ThronesSearchConditions
  queries!: ThronesSearchConditions
}
```

This doesn't yet add the `queries` property, but refactors to the
current implementation into `filters` in a reusable way.

One consequence of this is the existing syntax becomes odd:

```ts
search.filters.name.eq("foo").and.filters.title.eq("bar")
```

That `and` only has the possibility to add additional filters, not
queries. Plus it's repetitive. So the brunt of this code removes it:

```ts
search.filters.name.eq("foo").and.title.eq("bar")
````

We do this by (unfortunately but necessarily) faking the type. Though
we're returning an `AndClause` object, we're telling TS it's a
`Conditions` object:

```ts
get and(): ConditionsT & KeywordCondition<ConditionsT> {
  return applyAndClause(this, KeywordAndClause) as unknown as ConditionsT & KeywordCondition<ConditionsT>
}
```

Since this should only affect internal library authors, I think the good
far outweighs the bad.

The last bit here: since we're changing to `filters` instead of
`conditions`, let's go ahead and more closely match the JSON:API
payload. Might as well stick with a convention:

```json
{
  "filters": {...},
  "page": { "size": 10, "number": 2 },
  "sort": [{ "att": "name", "dir": "desc" }]
}
```